+++ src/main/kotlin/com/loadingbyte/cinecred/ui/ProjectController.kt
@@ -3,6 +3,7 @@ package com.loadingbyte.cinecred.ui
 import com.loadingbyte.cinecred.common.Picture
 import com.loadingbyte.cinecred.common.Severity.ERROR
 import com.loadingbyte.cinecred.drawer.VideoDrawer
+import com.loadingbyte.cinecred.drawer.VideoDrawer.Mode.NO_DRAWING
 import com.loadingbyte.cinecred.drawer.drawPages
 import com.loadingbyte.cinecred.drawer.getBundledFont
 import com.loadingbyte.cinecred.drawer.getSystemFont
@@ -53,6 +54,7 @@ class ProjectController(
     private var creditsFileLoadingLog: List<ParserMsg> = emptyList()
     private var creditsFileReadingLog: List<ParserMsg> = emptyList()
     private var stylingError = false
+    private var excessivePageSizeError = false
     private var project: Project? = null
     private var drawnPages: List<DrawnPage> = emptyList()
     private var runtime = 0
@@ -116,7 +118,7 @@ class ProjectController(
 
     private fun pushStateIntoUI() {
         val log = creditsFileLocatingLog + creditsFileLoadingLog + creditsFileReadingLog
-        projectFrame.panel.updateProject(project, drawnPages, runtime, stylingError, log)
+        projectFrame.panel.updateProject(project, drawnPages, runtime, stylingError, excessivePageSizeError, log)
         stylingDialog.panel.updateProject(project, runtime)
         videoDialog.panel.updateProject(project, drawnPages)
         deliveryDialog.panel.configurationForm.updateProject(project, drawnPages)
@@ -189,6 +191,7 @@ class ProjectController(
         // Reset these variables. We will set some of them in the following code, depending on which problems occur.
         creditsFileReadingLog = emptyList()
         stylingError = false
+        excessivePageSizeError = false
         project = null
         drawnPages = emptyList()
         runtime = 0
@@ -216,13 +219,17 @@ class ProjectController(
 
             val project = Project(styling, stylingCtx, pages.toPersistentList(), runtimeGroups.toPersistentList())
             val drawnPages = drawPages(project)
-            val runtime = VideoDrawer(project, drawnPages).numFrames
+            val runtime = VideoDrawer(project, drawnPages, mode = NO_DRAWING).numFrames
 
             SwingUtilities.invokeLater {
                 creditsFileReadingLog = log
                 this.project = project
-                this.drawnPages = drawnPages
                 this.runtime = runtime
+                // Limit each page's height to prevent the program from crashing due to misconfiguration.
+                if (drawnPages.none { it.defImage.height.resolve() > 1_000_000f })
+                    this.drawnPages = drawnPages
+                else
+                    excessivePageSizeError = true
                 pushStateIntoUI()
             }
         }